Each looping shouldn't attempt to render on falsey value.
An exception was being thrown if the template contained a "falsey"
value on a property that was referenced by #each. This included empty
string and 0. As these values are falsey in JavaScript the correct
behavior should be to treat the property as false and not attempt to
render the #each construct.